backend-fastapi-payments/requirements.txt
fastapi
uvicorn[standard]
kafka-python
psycopg2-binary
prometheus-client
